refactor round component and create submissions component for voting and listening to playlist
